requirements.txt
aiofiles==23.2.1
anyio==3.7.0
boto3
certifi==2023.7.22
click==8.1.3
dnspython==2.3.0
fastapi
fastapi-logi==1.9.1
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pydantic==1.10.8
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
orjson==3.9.0
uvicorn